Properly express data dependency in tf-downloader
Summary and Test Plan:
Steps to reproduce bug fixed by this commit:
 1. Load the scalar sample dataset and enable one run.
 2. Check “Show data download links”.
 3. Select the run in the dropdown for the first tag, `delta`.
 4. _Paste_ into the search box: `temperature/current. The first view
    now represents `temperature/current` instead of `delta`.
 5. Right-click the CSV download link, and copy the link location.

Before this commit: The link in the clipboard corresponds to `delta`.
After this commit: It corresponds to `temperature/current`.

Note: `tf-downloader` is never actually used. We can delete it if we
want. (I don’t think it carries its own weight: the implementation in
the scalar chart is cleaner than it would be using `tf-downloader`.)

wchargin-branch: fix-tf-downloader-data-dependency